feat: resolve projects
